Adding the Files with the Task and the Solution
The .pdf file contains the task itself (as published on HackerRank), and the .sql file contains my solution. I've decided to focus on the abilities of window functions while completing this task since it seems like a more efficient way to do so. Even though a more obvious way would be a recursion :)